// types.ts
export type Peli = {
  name: string;
  brand: string;
  iso: number;
  formatThirtyFive: boolean;
  formatOneTwenty: boolean;
  color: boolean;
  process: string;
  staticImageUrl: string;
  description: string;
  customDescription: any[];
  keyFeatures: { _id: string; feature: string }[];
  dateAdded: string;
  __v: number;
};